feat: move assets to separate package `@tosu/pages`, apply vite build system
* Vite build system for pages.
* Remove page build scripts and deps from `@tosu/server`.
* Use vue package.
