+++ packages/pages/homepage.html
@@ -2,10 +2,11 @@
 
 <head>
   <title>tosu dashboard v1.5</title>
-  <link rel="icon" type="image/x-icon" href="/assets/favicon.ico" />
-  <link rel="stylesheet" href="/assets/homepage.min.css" />
-  <link rel="stylesheet" href="/assets/icons/style.css" />
-  <link rel="stylesheet" href="/assets/fonts/style.css" />
+  <meta charset="UTF-8">
+  <link rel="icon" type="image/x-icon" href="./favicon.ico" />
+  <link rel="stylesheet" href="./src/homepage.scss" />
+  <link rel="stylesheet" href="./assets/icons/style.css" />
+  <link rel="stylesheet" href="./assets/fonts/style.css" />
 </head>
 
 <body>
@@ -77,8 +78,7 @@
       <a href="https://github.com/xxCherry" target="_blank">Cherry</a>
     </span>
   </footer>
-  <script src="/assets/vue.js"></script>
-  <script type="module" src="/assets/homepage.js"></script>
+  <script type="module" src="./src/homepage.js"></script>
 </body>
 
 </html>
