+++ packages/pages/ingame.html
@@ -2,16 +2,16 @@
 <html lang="en">
   <head>
     <title>tosu ingame</title>
-    <meta charset="UTF-8" />
+    <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
 
-    <link rel="stylesheet" href="/assets/ingame.css" />
-    <link rel="stylesheet" href="/assets/homepage.min.css" />
-    <link rel="stylesheet" href="/assets/icons/style.css" />
-    <link rel="stylesheet" href="/assets/fonts/style.css" />
+    <link rel="stylesheet" href="./src/ingame.css" />
+    <link rel="stylesheet" href="./src/homepage.scss" />
+    <link rel="stylesheet" href="./assets/icons/style.css" />
+    <link rel="stylesheet" href="./assets/fonts/style.css" />
 
-    <link rel="preload" as="image" href="/assets/images/39979.png" />
+    <link rel="preload" as="image" href="./assets/images/39979.png" />
   </head>
 
   <body class="ingame">
@@ -32,7 +32,7 @@
       <div class="scroll">
         <div v-if="available_overlays.length == 0" class="no-available">
           <span>You have 0 overlays</span><br />
-          <a href="http://127.0.0.1:24050/available?ingame=true" target="_blank">
+          <a href="/available?ingame=true" target="_blank">
             Download more
             <i class="icon-link"></i>
           </a>
@@ -63,7 +63,7 @@
           <a
             v-if="overlays[context_overlay.index]._settings == true"
             class="ctx-button btn-settings"
-            :href="`http://127.0.0.1:24050/settings?overlay=${overlays[context_overlay.index].folderName}&ingame=true`"
+            :href="`/settings?overlay=${overlays[context_overlay.index].folderName}&ingame=true`"
             target="_blank"
             @click.left="notification({
                 text: `Settings opened in separate window`,
@@ -216,7 +216,7 @@
         </div>
       </transition-group>
 
-      <a href="http://127.0.0.1:24050/available?ingame=true" target="_blank" class="ctx-button download-more">
+      <a href="/available?ingame=true" target="_blank" class="ctx-button download-more">
         <i class="icon-link"></i>
         Download more overlays
       </a>
@@ -245,8 +245,7 @@
 
     <div class="copyright">Tosu.app</div>
 
-    <script src="/assets/vue.js"></script>
-    <script type="module" src="/assets/ingame.js"></script>
-    <script type="module" src="/assets/homepage.js"></script>
+    <script type="module" src="./src/ingame.js"></script>
+    <script type="module" src="./src/homepage.js"></script>
   </body>
 </html>
