+++ packages/pages/package.json
@@ -1,0 +1,22 @@
+{
+  "name": "@tosu/pages",
+  "version": "0.1.0",
+  "description": "Web pages for tosu",
+  "author": "cyperdark",
+  "type": "module",
+  "exports": {
+    "./*": "./dist/*"
+  },
+  "scripts": {
+    "build": "vite build"
+  },
+  "dependencies": {
+    "vue": "^3.5.27"
+  },
+  "devDependencies": {
+    "@vitejs/plugin-vue": "^6.0.3",
+    "autoprefixer": "^10.4.23",
+    "vite": "^7.3.0",
+    "vite-tsconfig-paths": "^6.0.4"
+  }
+}
