+++ packages/pages/public/favicon.ico
