+++ packages/pages/report.html
@@ -3,7 +3,7 @@
 
 <head>
   <title>Tosu Report</title>
-  <meta charset="UTF-8" />
+  <meta charset="UTF-8">
   <meta http-equiv="X-UA-Compatible" content="IE=edge" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <style>
