--- packages/server/assets/icons/style.css
@@ -1,1 +1,0 @@
-@font-face{font-family:icomoon;src:url(fonts/icomoon.eot?yo4fz7);src:url(fonts/icomoon.eot?yo4fz7#iefix) format('embedded-opentype'),url(fonts/icomoon.ttf?yo4fz7) format('truetype'),url(fonts/icomoon.woff?yo4fz7) format('woff'),url(fonts/icomoon.svg?yo4fz7#icomoon) format('svg');font-weight:400;font-style:normal;font-display:block}[class*=" icon-"],[class^=icon-]{font-family:icomoon!important;speak:never;font-style:normal;font-weight:400;font-variant:normal;text-transform:none;line-height:1;-webkit-font-smoothing:antialiased;-moz-osx-font-smoothing:grayscale}.icon-plus:before{content:"\e92a"}.icon-delete:before{content:"\e92d"}.icon-folder:before{content:"\e900"}.icon-settings:before{content:"\e901"}.icon-link:before{content:"\e902"}.icon-github:before{content:"\e903"}.icon-discord:before{content:"\e904"}.icon-builder:before{content:"\e905"}
