+++ packages/server/package.json
@@ -4,9 +4,7 @@
   "description": "",
   "author": "cyperdark",
   "main": "index.ts",
-  "scripts": {
-    "prepare": "sass --style compressed --no-source-map ./assets/homepage.scss ./assets/homepage.min.css && postcss ./assets/homepage.min.css -o ./assets/homepage.min.css --use autoprefixer --no-map"
-  },
+  "scripts": {},
   "dependencies": {
     "@kotrikd/rosu-pp": "3.1.0",
     "@tosu/common": "workspace:*",
@@ -17,9 +15,6 @@
   },
   "devDependencies": {
     "@types/semver": "^7.7.1",
-    "@types/ws": "^8.18.1",
-    "autoprefixer": "^10.4.23",
-    "postcss-cli": "^11.0.1",
-    "sass": "^1.97.1"
+    "@types/ws": "^8.18.1"
   }
 }
