+++ packages/server/router/assets.ts
@@ -4,10 +4,7 @@ import path from 'path';
 
 import { Server, getContentType } from '../index';
 
-const pkgAssetsPath =
-    'pkg' in process
-        ? path.join(__dirname, 'assets')
-        : path.join(__dirname, '../assets');
+const pkgAssetsPath = path.join(__dirname, 'assets');
 
 export default function buildAssetsApi(server: Server) {
     server.app.route(/^\/assets\/(?<filePath>.*)/, 'GET', (req, res) => {
