+++ packages/server/utils/report.ts
@@ -97,10 +97,7 @@ export async function generateReport(instanceManager: any): Promise<Report> {
     };
 }
 
-const pkgAssetsPath =
-    'pkg' in process
-        ? path.join(__dirname, 'assets')
-        : path.join(__dirname, '../assets');
+const pkgAssetsPath = path.join(__dirname, 'assets');
 
 export async function generateReportHTML(report: Report): Promise<string> {
     const rawHtml = await readFile(
